+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds monitoring for APM trace errors.
+
 ## [v18.4.2](https://github.com/lexicalunit/spellbot/releases/tag/v18.4.2) - 2026-04-20
 
 ### Changed